admin/operations: revert 490a72b575c75d1aab103807ca7d67279a3b3706 and improve requeueing
Using a filename which cannot be interpreted as integer value in decimal notation
without any leading zeros does not work[1]. The **only** id which can be used save seems
to be `0`, other values can be overwritten due to [2] and the implementation
of `message_enqueue_retrieve_max_ID`[3] which is **not** free of race conditions. So
better be save than sorry.

[1] https://github.com/grommunio/gromox/issues/135
[2] https://github.com/grommunio/gromox/pull/136
[3] https://github.com/grommunio/gromox/blob/17ca866e731a3c19a6e66491442dd68a72b0ad39/mda/message_enqueue.cpp#L328

The code has been improved:
- always use `mv -n` to **never** replace existing files
- do not croak if `save/` is empty